`Main` split into four modules
patchnotes:
-main has been split into 4 modules
-`Engine.hs` for internal data types and variables
-'Logic.hs' for game logic/state updating
-'Draw.hs' for game render functions
-`Main.hs` for main game cycle
