+++ package.yaml
@@ -1,5 +1,5 @@
 name:                helloworld
-version:             1.0.3.0
+version:             1.0.3.1
 github:              "heightweight/pong"
 license:             BSD3
 author:              "Boiko Volodymyr"
