+++ src/Lib.hs
@@ -2,8 +2,14 @@ module Lib
     ( someFunc
     ) where
 
-import Graphics.Gloss
-import Graphics.Gloss.Interface.Pure.Game
+--import Graphics.Gloss
+--import Graphics.Gloss.Interface.Pure.Game
+--import Graphics.Gloss.Interface.IO.Game
+--import Control.Monad.Trans
+--import System.Directory
+--import Data.List.Split
+--import Data.List
+--import Data.Tuple
 
 someFunc :: IO ()
 someFunc = putStrLn "someFunc"
